requirements.txt
# Backend & API
fastapi
uvicorn[standard]
python-multipart

# Machine Learning & Data Processing
tensorflow==2.15.0
keras==2.15.0
opencv-python-headless<4.10
numpy<2.0.0
pandas
scikit-learn

# Visualization & Image Handling (Optional for API, but kept for Streamlit/Testing)
matplotlib
seaborn
Pillow
openpyxl

# Frontend (Keep if you are still running your Streamlit UI in the same environment)
streamlit